Tests/LibWeb/Text/expected/css/unknown-webkit-pseudo-element-serialization.txt

.d-none, .oops::-webkit-asdf
